FEATURE: health status and proper exit code for node:repair
This is a backwards compatible rework of the `node:repair` command.

This introduces a new `EventDispatchingNodeCommandControllerPluginInterface`
that can be implemented by NodeCommandControllerPlugins in order to
dispatch repair events instead of having to interact with the console
output directly.

The two provided core plugins now implement this interface and dispatch
corresponding events.

Background:
This rework allows the Command Controller to control the execution of tasks
and the rendering of their output which allows for a more streamlined
console output and serves as basis for further improvements (e.g. verbosity
flag).
This also fixes the exit code of the command, that is now 1 if one of the
commands triggered an error.

Resolves: #1988