!!! TASK: Disable and deprecate plugin and content prototype generators
The prototype generator classes are deprecated and removed from the nodeType configuration. This
helps to make the fusion-code more explict and understand the connection between nodeType and fusion
prototype much simpler. The previously assigned base types the were not recommended anymore and
thus usually are overwritten anyways.

A very welcome effect is that the error message for missing node rendering prototypes will say just that
and not any more that a template file is missing.

In addition this removes the documentPrototypeGenerator that was deprecated already for Neos 5.