BUGFIX: Fix broken JS require implementation in Backend Modules
With resolution of #2479 javascript blocks were put into CDATA
sections in order to prevent them to break the Fluid parser.

This introduced a regression that prevents Backend Modules to
work properly when `loadMinifiedJavascript` is disabled.

Background:

The string `resource://...` is replaced by the `ResourceInterceptor`
but that behavior is skipped for CDATA sections.

Related: #2479