FEATURE: Reduced stylesheet variant for backend modules
With this change you can tell Neos to load a reduced
stylesheet for your backend module that will provide the necessary styles and resets for backend modules that use Neos styles but had issues with the old „Full“ styles and resets. This stylesheet is also only about 50% of the old one.

You can switch to the reduced stylesheet in your configuration like this:
```
Neos:
  Neos:
    modules:
      user: # Or any other module group
        submodules:
          myModule:
            mainStylesheet: Lite
```