FEATURE: Allow users without access to the content module
Before, it was not possible to log in to the neos backend without having
access to the content module, as the login redirected to this module by default.

With this feature, an available module is determined when a user logs in and the user is redirected to that module.

There is a setting `Neos.Neos.moduleConfiguration.preferredStartModules` available which holds an array with preferred start modules. The array is matched with the available modules for the user and the first matching is chosen. If no matching module is found, the first modul accessible module is taken.